refactor(ReactWindows): Reverting changes outside ReactWindows folder
Starting to move towards a model where ReactWindows can be layered on top of React Native without requiring any core JS changes. The approach here is to take any of the files that had been modified for the Windows platform and copy them into the ReactWindows sub-folder. Since this obviously breaks the packager, as there would now be multiple modules with the same name, we place all the overridden modules in the rn-cli blacklist.